chore: move model proxy to separate pkg
The model proxy is currently housed within the rpc package.
Originally I did this because the model proxy relied on the Juju RPC message definition declared in the rpc package.
But this  is likely not the right place for the model proxy as it has nothing to do with the dialer logic in the rpc package.
Another supporting argument that the model proxy was in the wrong package is that the jimm package relies on the rpc package for dialing logic, but does not rely on the model proxy at all.
And in fact, instead, the model proxy indirectly relies on the jimm package for business logic.
