+++ internal/rpcproxy/message.go
@@ -1,20 +1,17 @@
 // Copyright 2025 Canonical.
 
-// Package rpc implements the juju RPC protocol. The main difference
-// between this implementation and the implementation in
-// github.com/juju/juju/rpc is that this implementation allows canceling
-// individual RPC calls using contexts. This implementation is a lot less
-// flexible than the juju implementation.
-package rpc
+package rpcproxy
 
 import (
 	"encoding/json"
+	"time"
 )
 
 // A message encodes a single message sent, or received, over an RPC
 // connection. It contains the union of fields in a request or response
 // message.
 type message struct {
+	start     time.Time
 	RequestID uint64                 `json:"request-id,omitempty"`
 	Type      string                 `json:"type,omitempty"`
 	Version   int                    `json:"version,omitempty"`
